Cleanup

+++ SecurityAssignment3/src/main/java/org/example/MessageGenerator.java
@@ -27,7 +27,7 @@ public class MessageGenerator {
         byte[] addressPadding = new byte[]{(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
                 (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0xFF, (byte) 0xFF};
         Random random = new Random();
-        long nonce = random.nextLong(Integer.MAX_VALUE);;
+        long nonce = random.nextLong((long) Integer.MAX_VALUE);;
         String userAgent = "/Satoshi:27.0.0/"; //Ignoring this as it's not needed, can be uncommented down below
         int startHeight = 0;
         boolean relay = true;
